% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.sum.R
\name{my.sum}
\alias{my.sum}
\title{Sum two numeric objects.}
\usage{
my.sum(a, b)
}
\arguments{
\item{a}{A real number}

\item{b}{A real number}
}
\value{
The sum of \code{a} and \code{b}
}
\description{
Because this is a very simple function, my explanation is short. These
paragraphs should explain everything you need to know.
}
\details{
This is still in the description part of the documentation and and it
will be until we see something that indicates a new section.
}
\examples{
sum(12,5)
sum(4,-2)
}
